is_float_column moved from the PropertyLabelCorrelation to the utils to avoid duplication
